# generated/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  license:
    name: KSTU
  title: Tourist agency
  version: 1.0.0
servers:
- url: http://127.0.0.1/v1
- url: https://fdobrotv.com/v1
paths:
...
  /foodOptions:
    get:
      operationId: listFoodOptions
      parameters:
      - description: How many items to return at one time (max 100)
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int32
          maximum: 100
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodOptions'
          description: A paged array of foodOptions
          headers:
            x-next:
              description: A link to the next page of responses
              explode: false
              schema:
                type: string
              style: simple
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List all foodOptions
      tags:
      - foodOptions
      x-accepts: application/json
      x-tags:
      - tag: foodOptions
    post:
      operationId: createFoodOption
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoodOptionIn'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodOption'
          description: Ok
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create a foodOption
      tags:
      - foodOptions
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: foodOptions
  /foodOptions/{id}:
    delete:
      operationId: deleteFoodOptionById
      parameters:
      - description: The id of the resource to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Resource deleted successfully
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Delete specific food option by id
      tags:
      - foodOptions
      x-accepts: application/json
      x-tags:
      - tag: foodOptions
    get:
      operationId: showFoodOptionById
      parameters:
      - description: The id of the resource to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodOption'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Info for a specific food option
      tags:
      - foodOptions
      x-accepts: application/json
      x-tags:
      - tag: foodOptions
...
components:
  schemas:
...
    FoodOptions:
      items:
        $ref: '#/components/schemas/FoodOption'
      maxItems: 100
      type: array
    FoodOptionIn:
      example:
        price: 1000
        name: All inclusive with alcohol
      properties:
        name:
          example: All inclusive with alcohol
          type: string
        price:
          example: 1000
          type: number
      required:
      - name
      - price
      type: object
    FoodOption:
      allOf:
      - $ref: '#/components/schemas/FoodOptionIn'
      - properties:
          id:
            format: uuid
            type: string
        required:
        - id
        type: object
    Error:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
      - code
      - message
      type: object